+++ .vscode-test.mjs
@@ -1,5 +1,8 @@
 import { defineConfig } from '@vscode/test-cli';
 
 export default defineConfig({
-	files: 'out/test/**/*.test.js',
+	files: [
+		'out/test/unit/**/*.test.js',
+		'out/test/integration/**/*.test.js',
+	],
 });